Create R markdown file
This R Markdown file presents a study on the effects of social media on communication and mental health using data analysis techniques.  It will perform filtering and cleaning of the data, which has been included as a csv file. 

The R program performs Correlational Analysis and a Regression analysis and visualizes the results using plots (regression plot and correlational matrix).